add word scramble app and add text files for the game
