+++ type_mania/templates/type_mania/type_mania.html
@@ -132,7 +132,7 @@
         <div class="center-items">
           <li class="recall"><a href="{% url 'recall_it' %}">Recall It</a></li>
           <li class="type-mania"><a href="{% url 'type_mania' %}">Type Mania</a></li>
-          <li><a href="wordScramble.html">Word Scramble</a></li>
+          <li><a href="{% url 'word_scramble' %}">Word Scramble</a></li>
         </div>
         <div class="right-items">
           {% if user.is_authenticated %}
