+++ word_scramble/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .views import show_game
+
+urlpatterns = [
+    path('', show_game, name='word_scramble')
+]
